move to unsigned fd for read() and write()

+++ lib/include/io.h
@@ -10,8 +10,8 @@
 #define STDOUT 1
 #define STDERR 2
 
-RESULT(u64) read(int fd, void* buffer, u64 count);
-RESULT(u64) write(int fd, void const* data, u64 count);
+RESULT(u64) read(u64 fd, void* buffer, u64 count);
+RESULT(u64) write(u64 fd, void const* data, u64 count);
 
 #define O_RDONLY    00000000
 #define O_WRONLY    00000001
